Fixed some 'debug' problems, the same problems exist in the trunk, but since I'm focusing on the scpu64 branch I'm doing the fixes in this branch, they will be merged into the trunk at a later time.

+++ branches/soci/vice-scpu64/src/debug.c
@@ -390,7 +390,7 @@ static int debug_autoplay_current_frame;
 static void debug_close_file(void)
 {
     if (debug_file != NULL) {
-        if (fwrite(debug_buffer, sizeof(char), debug_buffer_ptr, debug_file) < debug_buffer_ptr) {
+        if (fwrite(debug_buffer, sizeof(char), debug_buffer_ptr, debug_file) < (size_t)debug_buffer_ptr) {
             fprintf(stderr, "error writing debug log.\n");
         }
         fclose(debug_file);
